requirements.txt
duckduckgo-search>=3.0.0
RestrictedPython>=6.0
pandas>=1.5.0
numpy>=1.23.0
pydantic>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0 